+++ src/main/java/com/isul/controller/BoardController.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.isul.board.BoardService;
 import com.isul.dto.BoardDTO;
@@ -18,23 +17,15 @@ public class BoardController {
 	
 	@Autowired public BoardService boardService;
 	
-	/*
-	 * @RequestMapping("/csboard") public String csboardView() {
-	 * System.out.println("고객센터 열기"); return "board/csboard"; }
-	 */
-
-	
-	  
-	  
-	  @GetMapping("/csboard") 
-	  public String csBoardView(BoardDTO boardDTO, Model model){ 
-		  System.out.println("게시판 열기");
+	@GetMapping("/csboard") 
+	public String csBoardView(BoardDTO boardDTO, Model model){ 
+	  System.out.println("게시판 열기");
+	 
+	  List<BoardDTO> boardList = boardService.getBoardList(boardDTO);
+	  model.addAttribute("boardList", boardList);
+	  System.out.println("보드리스트 열기"+boardList);
 		 
-		  List<BoardDTO> boardList = boardService.getBoardList(boardDTO);
-		  model.addAttribute("boardList", boardList);
-		  System.out.println("보드리스트 열기"+boardList);
-			 
-		  return "board/csboard";
-	  }
+	  return "board/csboard";
+	}
 	 
 }
