+++ src/main/resources/mappers/MemberMapper.xml
@@ -50,7 +50,7 @@
 
 
 	<update id="changePassword" parameterType="com.isul.dto.MemberDTO">
-		UPDATE member SET password=#{password} WHERE email=#{email}
+		UPDATE member SET password=#{password} WHERE id=#{id} AND email=#{email}
 	</update>
 
 	<!-- 서로 친구 프로필목록 -->
