+++ src/main/resources/templates/index.html
@@ -65,7 +65,7 @@ a:link, a:visited, a:hover, a:active {
 <body>
 	<div class="background">
 		<div class="container">
-			<form action="/login" method="post">
+			<form action="/member/login" method="post">
 				<table style="margin:auto;">
 					<tr>
 						<td colspan="3">
@@ -85,7 +85,7 @@ a:link, a:visited, a:hover, a:active {
 					</tr>
 					<tr>
 						<td colspan="3">
-						<input type="submit" style="width:314px; height:36px;" onclick="return login()" value="로그인" class="login_form">
+						<input type="submit" style="width:314px; height:36px;" value="로그인" class="login_form">
 						</td>
 					</tr>
 					<tr>
